+++ packages/react-headless/slider/src/index.ts
@@ -1,0 +1,22 @@
+export {
+  SliderRoot,
+  SliderThumb,
+  SliderRange,
+  SliderHiddenInput,
+  SliderTick,
+  SliderMarker,
+  SliderValueIndicatorRoot,
+  SliderValueIndicatorLabel,
+  type SliderRootProps,
+  type SliderThumbProps,
+  type SliderRangeProps,
+  type SliderHiddenInputProps,
+  type SliderTickProps,
+  type SliderMarkerProps,
+  type SliderValueIndicatorRootProps,
+  type SliderValueIndicatorLabelProps,
+} from "./Slider";
+
+export * as Slider from "./Slider.namespace";
+
+export { useSliderContext } from "./useSliderContext";