+++ packages/react/src/components/HelpBubble/HelpBubble.tsx
@@ -13,7 +13,24 @@ const withStateProps = createWithStateProps([usePopoverContext]);
 
 ////////////////////////////////////////////////////////////////////////////////////
 
-export interface HelpBubbleRootProps extends HelpBubbleVariantProps, PopoverPrimitive.RootProps {}
+export interface HelpBubbleRootProps extends HelpBubbleVariantProps, PopoverPrimitive.RootProps {
+  /**
+   * @default "top"
+   */
+  placement?: PopoverPrimitive.RootProps["placement"];
+  /**
+   * @default 4
+   */
+  gutter?: PopoverPrimitive.RootProps["gutter"];
+  /**
+   * @default 16
+   */
+  overflowPadding?: PopoverPrimitive.RootProps["overflowPadding"];
+  /**
+   * @default 14
+   */
+  arrowPadding?: PopoverPrimitive.RootProps["arrowPadding"];
+}
 
 export const HelpBubbleRoot = withRootProvider<HelpBubbleRootProps>(PopoverPrimitive.Root, {
   defaultProps: {
@@ -21,9 +38,6 @@ export const HelpBubbleRoot = withRootProvider<HelpBubbleRootProps>(PopoverPrimi
     gutter: 4, // TODO: get value from rootage spec
     overflowPadding: 16,
     arrowPadding: 14,
-    flip: true,
-    slide: true,
-    strategy: "absolute",
   },
 });
 
@@ -74,7 +88,7 @@ export const HelpBubbleArrow = withContext<HTMLDivElement, HelpBubbleArrowProps>
 export interface HelpBubbleArrowTipProps extends React.SVGProps<SVGSVGElement> {
   /**
    * radius of the arrow tip
-   * @default 1
+   * @default 2
    */
   tipRadius?: number;
 }