+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @seed-design/mcp
 
+## 1.0.7
+
+### Patch Changes
+
+- @seed-design/figma@1.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [6aafce0]
+  - @seed-design/figma@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- @seed-design/figma@1.0.5
+
 ## 1.0.3
 
 ### Patch Changes