+++ packages/rootage/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @seed-design/rootage-artifacts
 
+## 1.0.6
+
+### Patch Changes
+
+- 15ab93a: List Item 컴포넌트의 상하 여백을 `$dimension.x2_5`에서 `$dimension.x3`로 늘립니다.
+
+## 1.0.5
+
+### Patch Changes
+
+- 6aafce0: Tag Group 컴포넌트를 추가합니다. Tag Group은 아이콘 및 텍스트로 이루어진 태그를 구분 기호와 함께 수평 레이아웃으로 표시하는 컴포넌트입니다.
+- 1902dfa: AppBar의 스타일을 업데이트합니다.
+
+  - Top Navigation의 title 및 description에 `lineHeight` 값을 정의합니다. (React AppBar 컴포넌트에 반영됩니다.)
+  - `<Icon />` 컴포넌트를 활용하여 `<AppBarIconButton />` 내부 아이콘을 커스터마이징할 수 있도록 수정합니다.
+
+- 4c33f07: Switch가 checked 상태가 아닐 때 thumb 크기를 줄여 상태를 인지하기 쉽도록 합니다. enabled-disabled 상태 간 트랜지션을 추가합니다.
+- 3df657f: Switch와 Switch Mark의 disabled 상태를 더 잘 구별할 수 있도록 스타일을 수정합니다.
+
+## 1.0.4
+
+### Patch Changes
+
+- f1cf4cd: Text Field와 Multiline Text Field가 기본적으로 배경 색을 갖지 않도록 수정합니다.
+- 3898183: 매너온도 L9 전경 및 배경 색상의 채도를 낮춥니다.
+
 ## 1.0.3
 
 ### Patch Changes